Mu.GRpc.Server

Description

This module allows you to server a Mu Server as a WAI Application using gRPC as transport layer.

The simples way is to use runGRpcApp, all other variants provide more control over the settings.

Supported messaging formats

data GRpcMessageProtocol

Constructors

MsgProtoBuf

MsgAvro

msgProtoBuf :: Proxy 'MsgProtoBuf

msgAvro :: Proxy 'MsgAvro

Run a Server directly

runGRpcApp :: (KnownName name, GRpcServiceHandlers ('Package ('Just name) services) protocol ServerErrorIO chn services handlers) => Proxy protocol -> Port -> ServerT chn () ('Package ('Just name) services) ServerErrorIO handlers -> IO ()

Run a Mu Server on the given port.

runGRpcAppTrans :: (KnownName name, GRpcServiceHandlers ('Package ('Just name) services) protocol m chn services handlers) => Proxy protocol -> Port -> (forall a. m a -> ServerErrorIO a) -> ServerT chn () ('Package ('Just name) services) m handlers -> IO ()

Run a Mu Server on the given port.

runGRpcAppSettings :: (KnownName name, GRpcServiceHandlers ('Package ('Just name) services) protocol m chn services handlers) => Proxy protocol -> Settings -> (forall a. m a -> ServerErrorIO a) -> ServerT chn () ('Package ('Just name) services) m handlers -> IO ()

Run a Mu Server using the given Settings.

Go to Warp to declare Settings.

data Settings

runGRpcAppTLS :: (KnownName name, GRpcServiceHandlers ('Package ('Just name) services) protocol m chn services handlers) => Proxy protocol -> TLSSettings -> Settings -> (forall a. m a -> ServerErrorIO a) -> ServerT chn () ('Package ('Just name) services) m handlers -> IO ()

Run a Mu Server using the given TLSSettings and Settings.

Go to WarpTLS to declare TLSSettings and to Warp to declare Settings.

data TLSSettings

Convert a Server into a WAI application

gRpcApp :: (KnownName name, GRpcServiceHandlers ('Package ('Just name) services) protocol ServerErrorIO chn services handlers) => Proxy protocol -> ServerT chn () ('Package ('Just name) services) ServerErrorIO handlers -> Application

Turn a Mu Server into a WAI Application.

These Applications can be later combined using, for example, wai-routes, or you can add middleware from wai-extra, among others.

gRpcAppTrans :: (KnownName name, GRpcServiceHandlers ('Package ('Just name) services) protocol m chn services handlers) => Proxy protocol -> (forall a. m a -> ServerErrorIO a) -> ServerT chn () ('Package ('Just name) services) m handlers -> Application

Turn a Mu Server into a WAI Application.

These Applications can be later combined using, for example, wai-routes, or you can add middleware from wai-extra, among others.

data WrappedServer protocol m where

Constructors

Srv :: (KnownName name, GRpcServiceHandlers ('Package ('Just name) services) protocol m chn services handlers) => ServerT chn () ('Package ('Just name) services) m handlers -> WrappedServer protocol m

gRpcMultipleApp :: Proxy protocol -> [WrappedServer protocol ServerErrorIO] -> Application

Turn several Mu Servers into a WAI Application.

These Applications can be later combined using, for example, wai-routes, or you can add middleware from wai-extra, among others.

gRpcMultipleAppTrans :: Proxy protocol -> (forall a. m a -> ServerErrorIO a) -> [WrappedServer protocol m] -> Application

Turn several Mu Servers into a WAI Application.

These Applications can be later combined using, for example, wai-routes, or you can add middleware from wai-extra, among others.

Raise errors as exceptions in IO

raiseErrors :: MonadIO m => ServerErrorIO a -> m a

Raises errors from ServerErrorIO as exceptions in a monad which supports IO.

This function is useful to interoperate with other libraries which cannot handle the additional error layer. In particular, with Conduit, as witnessed by liftServerConduit.

liftServerConduit :: MonadIO m => ConduitT a b ServerErrorIO r -> ConduitT a b m r

Turns a Conduit working on ServerErrorIO into any other base monad which supports IO, by raising any error as an exception.

This function is useful to interoperate with libraries which generate Conduits with other base monads, such as persistent.

Re-export useful instances